Update Vector3 to use floats

+++ psy/psy-vector3-private.h
@@ -3,8 +3,8 @@
 
 #include "psy-vector3.h"
 
-glm::dvec3&
+glm::vec3&
 psy_vector3_get_priv_reference(PsyVector3 *self);
 
-glm::dvec3 *
+glm::vec3 *
 psy_vector3_get_priv_pointer(PsyVector3 *self);
